refactor(app): extract RunFree into pkg/app, slim main.go to 201 lines (NVR-004)
Move NewApp/Start/Stop/buildRouter logic from cmd/mibee-nvr/main.go
into pkg/app/run.go as RunFree(cfg, configPath), which constructs all
managers and registers them as pkg/app.Service on a new *App
(start = registration order, stop = reverse order).

Key design: relay folded into the camera Service (relay starts before
camera, stops before camera) — preserves the asymmetry where relay
must be ready before camMgr.Start (SetCameraTargets) yet stop while
hubs still exist. Effective start order unchanged: relay, camera,
health, merge, transcode, mergeScheduler, cleanup, mqtt, ftp, rtmp, srt.

main.go: 1324 -> 201 lines. CLI subcommands extracted to cli.go.
HTTP server + signal handling stay in main.

Validation: go build/vet/test-race (48 passed) green; ARM64
cross-compile OK; binary starts, /api/health returns 200, SIGTERM
yields clean reverse-order shutdown + "MiBee NVR stopped" (exit 0).
